moved well_geometry to protcol_api core folder
