+++ api/src/opentrons/protocol_api/core/protocol_api/well.py
@@ -3,9 +3,9 @@ from typing import Optional
 
 from opentrons_shared_data.labware.constants import WELL_NAME_PATTERN
 
-from opentrons.protocols.geometry.well_geometry import WellGeometry
 from opentrons.types import Point
 
+from .well_geometry import WellGeometry
 from ..well import AbstractWellCore
