+++ api/tests/opentrons/protocol_api_old/test_labware.py
@@ -13,12 +13,12 @@ from opentrons.hardware_control.modules.types import (
 )
 
 from opentrons.protocols.api_support.types import APIVersion
-from opentrons.protocols.geometry.well_geometry import WellGeometry
 from opentrons.protocol_api import MAX_SUPPORTED_VERSION, labware, validation
 from opentrons.protocol_api.core.labware import AbstractLabware
 from opentrons.protocol_api.core.protocol_api import module_geometry
 from opentrons.protocol_api.core.protocol_api.labware import LabwareImplementation
 from opentrons.protocol_api.core.protocol_api.well import WellImplementation
+from opentrons.protocol_api.core.protocol_api.well_geometry import WellGeometry
 
 from opentrons.calibration_storage import helpers
 from opentrons.types import Point, Location
