+++ api/tests/opentrons/protocols/execution/test_execute_json_v3.py
@@ -1,6 +1,6 @@
-from opentrons.protocols.geometry.well_geometry import WellGeometry
 from opentrons.protocol_api.core.protocol_api.labware import LabwareImplementation
 from opentrons.protocol_api.core.protocol_api.well import WellImplementation
+from opentrons.protocol_api.core.protocol_api.well_geometry import WellGeometry
 
 from unittest import mock
 from copy import deepcopy
